제목 테그 적용, 배열 PUSH 제거, useState 나누기
