+++ babel.config.js
@@ -8,6 +8,8 @@ module.exports = {
         },
       },
     ],
-    '@babel/preset-react',
+    ["@babel/preset-react", {
+      "runtime": "automatic"
+    }]
   ],
 };
